# feature: Buffer.lastIndexOf