… src/string_search.h