@mikeal 2016-01-25
Oh wow, I've been waiting years for this :)

This should make writing parsers a lot nicer :)