@dcposch 2016-01-27
Also, this PR uses `memrchr` to search for a single byte value in a Buffer, back to front. Like `memchr`, it's probably a lot faster than just looping over bytes.

http://cebka.blogspot.com/2015/04/how-fast-is-your-memchr.html

**Unfortunately `memrchr` is GNU extension, not part of POSIX like `memchr`.**

1. Does Node have to compile in places where memrchr isn't available?
2. If so, should I try to pull in a polyfill, for example this one?
https://github.com/c9/node-gnu-tools/blob/master/grep-src/lib/memrchr.c