@feross 2016-01-28
Thanks for tackling this PR, @dcposch!