@dcposch 2016-01-29
@trevnorris thx. Fixed all the things you pointed out so far