@trevnorris 2016-02-01
@dcposch Left few more comments. Now that the weekend is over will be more attentive. :)