@dcposch 2016-02-05
I did some benchmarking. Summary: https://docs.google.com/spreadsheets/d/1Lpxiagv8EnCS9B5F6ed0Z8cVJn91vej0M-OkF9_Qg8g/edit#gid=0
Click on the sheet tab to see comments.

**TLDR; looks like I added a constant overhead of about ~85 clock cycles per call to indexOf().**

The benchmark vary: some of them are equally fast before and after this PR, and others got slower. Looking closer, the heavy expensive calls to indexOf() are equally fast and the quick calls to indexOf() got noticeably slower. Running just the worst result (28% slow down): 

**Before (795eeb1): ~9.3m ops / second**
At 2.2GHz = ~235 clock cycles per indexOf()
```
$ ./node benchmark/buffers/buffer-indexof.js type=buffer encoding=undefined search=Alice
buffers/buffer-indexof.js search=Alice encoding=undefined type=buffer iter=1: 9249264.91467
$ ./node benchmark/buffers/buffer-indexof.js type=buffer encoding=undefined search=Alice
buffers/buffer-indexof.js search=Alice encoding=undefined type=buffer iter=1: 9303733.69067
```

**After (b67eb8c): ~6.8m ops / second**
At 2.2GHz = ~320 clock cycles per indexOf()
```
$ ./node benchmark/buffers/buffer-indexof.js type=buffer encoding=undefined search=Alice
buffers/buffer-indexof.js search=Alice encoding=undefined type=buffer iter=1: 6886913.37014
$ ./node benchmark/buffers/buffer-indexof.js type=buffer encoding=undefined search=Alice
buffers/buffer-indexof.js search=Alice encoding=undefined type=buffer iter=1: 6783969.18233
```

So we added about 85 cycles overhead for every call to indexOf(). How many of those are in JS vs in C++?
Stubbing out the C++ part in `node_buffer.cc` like this:
```c++
void IndexOfBuffer(const FunctionCallbackInfo<Value>& args) {
  if (true) return args.GetReturnValue().Set(0);
```

Before (JS only, stubbed C++): 120 cycles per call
After (JS only, stubbed C++): 145 cycles per call

#### Summary

We added ~85 cycles of overhead to every call to indexOf(). About 25 of those are in the JS wrapper and the other 60 are in C++.

I think the overhead comes from:
* The new function in JS, `bidirectionalIndexOf`
* The extra `direction` flag in JS and several of the C++ functions
* The couple of branches that check the direction flag

I don't know how we'd optimize the JS part, short of copy pasting `bidirectionalIndexOf` into `indexOf` and `lastIndexOf`. For the C++ part, I guess we could turn the `dir` into a template argument instead of a function argument so it makes separate object code for `indexOf` vs `lastIndexOf`.

Dunno if we care.